+++ rust/src/nasl/inter/mod.rs
@@ -1,3 +1,5 @@
+mod error;
+
 mod assign;
 mod call;
 mod declare;
@@ -7,4 +9,9 @@ mod interpreter;
 mod loop_extension;
 mod operator;
 
+pub use error::{FunctionCallError, InterpretError, InterpretErrorKind};
 pub use forking_interpreter::ForkingInterpreter;
+pub use interpreter::Interpreter;
+
+#[cfg(test)]
+mod tests;
