+++ rust/src/nasl/inter/tests/local_var.rs
